requirements.txt
attr==0.3.2
brotli==1.2.0
brotlicffi==1.2.0.0
chardet==6.0.0.post1
ConfigParser==7.2.0
cryptography==46.0.5
docutils==0.22.4
fastapi==0.131.0
filelock==3.24.3
h2==4.3.0
HTMLParser==0.0.2
ipython==9.10.0
ipywidgets==8.1.8
jnius==1.1.0
js==1.0
keyring==25.7.0
pandas==3.0.1
protobuf==6.33.5
py_clob_client==0.34.6
pyodide==0.0.2
pyOpenSSL==25.3.0
python_dateutil==2.9.0.post0
pytz==2025.2
redis==7.2.0
simplejson==3.20.2
Sphinx==9.1.0
thread==2.0.6
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
uvicorn==0.41.0
websocket_client==1.9.0
xmlrpclib==1.0.1
